+++ pkg/exporter/sumologicexporter/graphite_formatter_test.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/model/pdata"
 )
 
 func TestEscapeGraphiteString(t *testing.T) {
@@ -71,162 +72,162 @@ func TestGraphiteFieldInvalidCharactersInValue(t *testing.T) {
 	assert.Equal(t, expected, result)
 }
 
-// func TestGraphiteMetricDataTypeIntGauge(t *testing.T) {
-// 	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
-// 	require.NoError(t, err)
+func TestGraphiteMetricDataTypeIntGauge(t *testing.T) {
+	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
+	require.NoError(t, err)
 
-// 	metric := exampleIntGaugeMetric()
-// 	metric.attributes.InsertString("cluster", "my_cluster")
-// 	metric.attributes.InsertString("namespace", "default")
-// 	metric.attributes.InsertString("pod", "some pod")
+	metric, attributes := exampleIntGaugeMetric()
+	attributes.InsertString("cluster", "my_cluster")
+	attributes.InsertString("namespace", "default")
+	attributes.InsertString("pod", "some pod")
 
-// 	result := gf.metric2String(metric)
-// 	expected := `my_cluster.default.some_pod.gauge_metric_name 124 1608124661
-// my_cluster.default.some_pod.gauge_metric_name 245 1608124662`
-// 	assert.Equal(t, expected, result)
-// }
+	result := gf.metric2String(metric, attributes)
+	expected := `my_cluster.default.some_pod.gauge_metric_name 124 1608124661
+my_cluster.default.some_pod.gauge_metric_name 245 1608124662`
+	assert.Equal(t, expected, result)
+}
 
-// func TestGraphiteMetricDataTypeDoubleGauge(t *testing.T) {
-// 	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
-// 	require.NoError(t, err)
+func TestGraphiteMetricDataTypeDoubleGauge(t *testing.T) {
+	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
+	require.NoError(t, err)
 
-// 	metric := exampleDoubleGaugeMetric()
-// 	metric.attributes.InsertString("cluster", "my_cluster")
-// 	metric.attributes.InsertString("namespace", "default")
-// 	metric.attributes.InsertString("pod", "some pod")
+	metric, attributes := exampleDoubleGaugeMetric()
+	attributes.InsertString("cluster", "my_cluster")
+	attributes.InsertString("namespace", "default")
+	attributes.InsertString("pod", "some pod")
 
-// 	result := gf.metric2String(metric)
-// 	expected := `my_cluster.default.some_pod.gauge_metric_name_double_test 33.4 1608124661
-// my_cluster.default.some_pod.gauge_metric_name_double_test 56.8 1608124662`
-// 	assert.Equal(t, expected, result)
-// }
+	result := gf.metric2String(metric, attributes)
+	expected := `my_cluster.default.some_pod.gauge_metric_name_double_test 33.4 1608124661
+my_cluster.default.some_pod.gauge_metric_name_double_test 56.8 1608124662`
+	assert.Equal(t, expected, result)
+}
 
-// func TestGraphiteNoattribute(t *testing.T) {
-// 	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
-// 	require.NoError(t, err)
+func TestGraphiteNoattribute(t *testing.T) {
+	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
+	require.NoError(t, err)
 
-// 	metric := exampleDoubleGaugeMetric()
-// 	metric.attributes.InsertString("cluster", "my_cluster")
-// 	metric.attributes.InsertString("pod", "some pod")
+	metric, attributes := exampleDoubleGaugeMetric()
+	attributes.InsertString("cluster", "my_cluster")
+	attributes.InsertString("pod", "some pod")
 
-// 	result := gf.metric2String(metric)
-// 	expected := `my_cluster..some_pod.gauge_metric_name_double_test 33.4 1608124661
-// my_cluster..some_pod.gauge_metric_name_double_test 56.8 1608124662`
-// 	assert.Equal(t, expected, result)
-// }
+	result := gf.metric2String(metric, attributes)
+	expected := `my_cluster..some_pod.gauge_metric_name_double_test 33.4 1608124661
+my_cluster..some_pod.gauge_metric_name_double_test 56.8 1608124662`
+	assert.Equal(t, expected, result)
+}
 
-// func TestGraphiteMetricDataTypeIntSum(t *testing.T) {
-// 	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
-// 	require.NoError(t, err)
+func TestGraphiteMetricDataTypeIntSum(t *testing.T) {
+	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
+	require.NoError(t, err)
 
-// 	metric := exampleIntSumMetric()
-// 	metric.attributes.InsertString("cluster", "my_cluster")
-// 	metric.attributes.InsertString("namespace", "default")
-// 	metric.attributes.InsertString("pod", "some pod")
+	metric, attributes := exampleIntSumMetric()
+	attributes.InsertString("cluster", "my_cluster")
+	attributes.InsertString("namespace", "default")
+	attributes.InsertString("pod", "some pod")
 
-// 	result := gf.metric2String(metric)
-// 	expected := `my_cluster.default.some_pod.sum_metric_int_test 45 1608124444
-// my_cluster.default.some_pod.sum_metric_int_test 1238 1608124699`
-// 	assert.Equal(t, expected, result)
-// }
+	result := gf.metric2String(metric, attributes)
+	expected := `my_cluster.default.some_pod.sum_metric_int_test 45 1608124444
+my_cluster.default.some_pod.sum_metric_int_test 1238 1608124699`
+	assert.Equal(t, expected, result)
+}
 
-// func TestGraphiteMetricDataTypeDoubleSum(t *testing.T) {
-// 	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
-// 	require.NoError(t, err)
+func TestGraphiteMetricDataTypeDoubleSum(t *testing.T) {
+	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
+	require.NoError(t, err)
 
-// 	metric := exampleDoubleSumMetric()
-// 	metric.attributes.InsertString("cluster", "my_cluster")
-// 	metric.attributes.InsertString("namespace", "default")
-// 	metric.attributes.InsertString("pod", "some pod")
+	metric, attributes := exampleDoubleSumMetric()
+	attributes.InsertString("cluster", "my_cluster")
+	attributes.InsertString("namespace", "default")
+	attributes.InsertString("pod", "some pod")
 
-// 	result := gf.metric2String(metric)
-// 	expected := `my_cluster.default.some_pod.sum_metric_double_test 45.6 1618124444
-// my_cluster.default.some_pod.sum_metric_double_test 1238.1 1608424699`
-// 	assert.Equal(t, expected, result)
-// }
+	result := gf.metric2String(metric, attributes)
+	expected := `my_cluster.default.some_pod.sum_metric_double_test 45.6 1618124444
+my_cluster.default.some_pod.sum_metric_double_test 1238.1 1608424699`
+	assert.Equal(t, expected, result)
+}
 
-// func TestGraphiteMetricDataTypeSummary(t *testing.T) {
-// 	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
-// 	require.NoError(t, err)
+func TestGraphiteMetricDataTypeSummary(t *testing.T) {
+	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
+	require.NoError(t, err)
 
-// 	metric := exampleSummaryMetric()
-// 	metric.attributes.InsertString("cluster", "my_cluster")
-// 	metric.attributes.InsertString("namespace", "default")
-// 	metric.attributes.InsertString("pod", "some pod")
+	metric, attributes := exampleSummaryMetric()
+	attributes.InsertString("cluster", "my_cluster")
+	attributes.InsertString("namespace", "default")
+	attributes.InsertString("pod", "some pod")
 
-// 	result := gf.metric2String(metric)
-// 	expected := ``
-// 	assert.Equal(t, expected, result)
-// }
+	result := gf.metric2String(metric, attributes)
+	expected := ``
+	assert.Equal(t, expected, result)
+}
 
-// func TestGraphiteMetricDataTypeHistogram(t *testing.T) {
-// 	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
-// 	require.NoError(t, err)
+func TestGraphiteMetricDataTypeHistogram(t *testing.T) {
+	gf, err := newGraphiteFormatter("%{cluster}.%{namespace}.%{pod}.%{_metric_}")
+	require.NoError(t, err)
 
-// 	metric := exampleHistogramMetric()
-// 	metric.attributes.InsertString("cluster", "my_cluster")
-// 	metric.attributes.InsertString("namespace", "default")
-// 	metric.attributes.InsertString("pod", "some pod")
+	metric, attributes := exampleHistogramMetric()
+	attributes.InsertString("cluster", "my_cluster")
+	attributes.InsertString("namespace", "default")
+	attributes.InsertString("pod", "some pod")
 
-// 	result := gf.metric2String(metric)
-// 	expected := ``
-// 	assert.Equal(t, expected, result)
-// }
+	result := gf.metric2String(metric, attributes)
+	expected := ``
+	assert.Equal(t, expected, result)
+}
 
-// func TestGraphiteMetrics(t *testing.T) {
-// 	type testCase struct {
-// 		name      string
-// 		metric    metricPair
-// 		expected  string
-// 		formatter string
-// 	}
+func TestGraphiteMetrics(t *testing.T) {
+	type testCase struct {
+		name       string
+		metricFunc func(fillData bool) (pdata.Metric, pdata.Map)
+		expected   string
+		formatter  string
+	}
 
-// 	tests := []testCase{
-// 		{
-// 			name:      "empty int gauge",
-// 			metric:    buildExampleIntGaugeMetric(false),
-// 			expected:  "",
-// 			formatter: "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
-// 		},
-// 		{
-// 			name:      "empty double gauge",
-// 			metric:    buildExampleDoubleGaugeMetric(false),
-// 			expected:  "",
-// 			formatter: "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
-// 		},
-// 		{
-// 			name:      "empty int sum",
-// 			metric:    buildExampleIntSumMetric(false),
-// 			expected:  "",
-// 			formatter: "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
-// 		},
-// 		{
-// 			name:      "empty double sum",
-// 			metric:    buildExampleDoubleSumMetric(false),
-// 			expected:  "",
-// 			formatter: "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
-// 		},
-// 		{
-// 			name:      "empty summary",
-// 			metric:    buildExampleSummaryMetric(false),
-// 			expected:  "",
-// 			formatter: "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
-// 		},
-// 		{
-// 			name:      "empty histogram",
-// 			metric:    buildExampleHistogramMetric(false),
-// 			expected:  "",
-// 			formatter: "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
-// 		},
-// 	}
+	tests := []testCase{
+		{
+			name:       "empty int gauge",
+			metricFunc: buildExampleIntGaugeMetric,
+			expected:   "",
+			formatter:  "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
+		},
+		{
+			name:       "empty double gauge",
+			metricFunc: buildExampleDoubleGaugeMetric,
+			expected:   "",
+			formatter:  "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
+		},
+		{
+			name:       "empty int sum",
+			metricFunc: buildExampleIntSumMetric,
+			expected:   "",
+			formatter:  "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
+		},
+		{
+			name:       "empty double sum",
+			metricFunc: buildExampleDoubleSumMetric,
+			expected:   "",
+			formatter:  "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
+		},
+		{
+			name:       "empty summary",
+			metricFunc: buildExampleSummaryMetric,
+			expected:   "",
+			formatter:  "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
+		},
+		{
+			name:       "empty histogram",
+			metricFunc: buildExampleHistogramMetric,
+			expected:   "",
+			formatter:  "%{cluster}.%{namespace}.%{pod}.%{_metric_}",
+		},
+	}
 
-// 	for _, tt := range tests {
-// 		t.Run(tt.name, func(t *testing.T) {
-// 			gf, err := newGraphiteFormatter(tt.formatter)
-// 			require.NoError(t, err)
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			gf, err := newGraphiteFormatter(tt.formatter)
+			require.NoError(t, err)
 
-// 			result := gf.metric2String(tt.metric)
-// 			assert.Equal(t, tt.expected, result)
-// 		})
-// 	}
-// }
+			result := gf.metric2String(tt.metricFunc(false))
+			assert.Equal(t, tt.expected, result)
+		})
+	}
+}